ml_reclamos/requirements.txt
-e ./ml_core
Flask
requests
python-dotenv
qrcode
reportlab
pillow
gunicorn

